Code Intergration and Changes
Took the code from guess and shuffle and put it into Make Puzzle.

Added code to New Puzzle check if the given baseword is in the database
